requirements.txt
Django
python-dotenv
black
flake8
isort
pre-commit
django-jazzmin
django-modeltranslation
django-phonenumber-field
djangorestframework
drf-yasg
psycopg2-binary
django-rosetta
django-ckeditor
djangorestframework-simplejwt
